+++ internal/util.go
@@ -14,11 +14,16 @@
 package internal
 
 import (
+	"context"
 	"fmt"
+	"io/ioutil"
 	"math"
+	"net/http"
 	"strconv"
 	"strings"
 
+	"github.com/go-kit/kit/log"
+	"github.com/go-kit/kit/log/level"
 	"github.com/kawamuray/jsonpath"
 	"github.com/prometheus-community/json_exporter/config"
 	"github.com/prometheus/client_golang/prometheus"
@@ -99,4 +104,33 @@ func CreateMetricsList(r *prometheus.Registry, c config.Config) ([]JsonGaugeColl
 		}
 	}
 	return metrics, nil
+}
+
+func FetchJson(ctx context.Context, logger log.Logger, endpoint string, headers map[string]string) ([]byte, error) {
+	client := &http.Client{}
+	req, err := http.NewRequest("GET", endpoint, nil)
+	req = req.WithContext(ctx)
+	if err != nil {
+		level.Error(logger).Log("msg", "Failed to create request", "err", err) //nolint:errcheck
+		return nil, err
+	}
+
+	for key, value := range headers {
+		req.Header.Add(key, value)
+	}
+	if req.Header.Get("Accept") == "" {
+		req.Header.Add("Accept", "application/json")
+	}
+	resp, err := client.Do(req)
+	if err != nil {
+		return nil, fmt.Errorf("failed to fetch json from endpoint;endpoint:<%s>,err:<%s>", endpoint, err)
+	}
+	defer resp.Body.Close()
+
+	data, err := ioutil.ReadAll(resp.Body)
+	if err != nil {
+		return nil, fmt.Errorf("failed to read response body;err:<%s>", err)
+	}
+
+	return data, nil
 }
